[JIMT][CT-884][CT-822 (partial)] generic useFileUploader (#61921)
* jimt/CT-822 - generic lab2 useFileUploader hook

* jimt/CT-822 - multi-file support; removes newFile FileId artefact

* jimt/CT-822 - added generic hook I forgot + updated tests

* jimt/CT-822 - brought newFolder into line with newFile w/o externally generated IDs

* jimt/CT-822 - removed erroneous debugging line

* jimt/CT-822 - added UPLOAD_SUCCEEDED generic analytic event

* jimt/CT-884 - fixed file upload bug

* jimt/CT-822 - incorpoate validationFileId into getNextFileId
