+++ apps/src/codebridge/codebridgeContext/types.ts
@@ -8,7 +8,6 @@ export type CloseFileFunction = (fileId: FileId) => void;
 export type SetActiveFileFunction = (fileId: FileId) => void;
 
 export type NewFolderFunction = (arg: {
-  folderId: FolderId;
   folderName: string;
   parentId?: FolderId;
 }) => void;
@@ -17,10 +16,10 @@ export type DeleteFolderFunction = (folderId: FolderId) => void;
 export type OpenFileFunction = (fileId: FileId) => void;
 export type DeleteFileFunction = (fileId: FileId) => void;
 export type NewFileFunction = (arg: {
-  fileId: FileId;
   fileName: string;
   folderId?: FolderId;
   contents?: string;
+  validationFileId?: string;
 }) => void;
 export type RenameFileFunction = (fileId: FileId, newName: string) => void;
 export type RenameFolderFunction = (folderId: string, newName: string) => void;
