+++ apps/src/codebridge/codebridgeContext/utils.ts
@@ -32,8 +32,12 @@ import {
   SetFileTypeFunction,
 } from './types';
 
-export const getNextFileId = (files: ProjectFile[]) => {
-  return String(Math.max(0, ...files.map(f => Number(f.id))) + 1);
+// optionally, we can hand in the validationFileId as the second argument. If it's included, then we'll use that as well
+// when we figure out the next file id. If we're not given one, then just set it to '0' so it doesn't interfere with id generation.
+export const getNextFileId = (files: ProjectFile[], validationFileId = '0') => {
+  return String(
+    Math.max(0, Number(validationFileId), ...files.map(f => Number(f.id))) + 1
+  );
 };
 
 export const getNextFolderId = (folders: ProjectFolder[]) => {
@@ -76,7 +80,6 @@ export const useProjectUtilities = (
         });
       }),
       newFile: <NewFileFunction>(({
-        fileId,
         fileName,
         folderId = DEFAULT_FOLDER_ID,
         // this line causes the apparent circular dependency issue
@@ -86,7 +89,6 @@ export const useProjectUtilities = (
         dispatch({
           type: PROJECT_REDUCER_ACTIONS.NEW_FILE,
           payload: {
-            fileId,
             fileName,
             folderId,
             contents: contents.replace(/\${fileName}/g, fileName),
@@ -150,13 +152,12 @@ export const useProjectUtilities = (
       }),
 
       newFolder: <NewFolderFunction>(({
-        folderId,
         folderName,
         parentId = DEFAULT_FOLDER_ID,
       }) => {
         dispatch({
           type: PROJECT_REDUCER_ACTIONS.NEW_FOLDER,
-          payload: {folderId, folderName, parentId},
+          payload: {folderName, parentId},
         });
       }),
       renameFolder: <RenameFolderFunction>((folderId, newName) => {
