+++ apps/src/lab2/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useFileUploader';
 export * from './useKeyboardTrap';
 export * from './useLifecycleNotifier';
 export * from './usePartialApply';
