+++ apps/test/unit/codebridge/FileBrowser/prompts/openNewFilePromptTest.ts
@@ -10,8 +10,7 @@ import {getDialogControlMock, getAnalyticsMock} from '../../test_utils';
 
 const getNewFileMock = (parentId: FolderId): [ProjectFile, NewFileFunction] => {
   const newFileData = {} as ProjectFile;
-  const mock: NewFileFunction = ({fileId, fileName, folderId}) => {
-    newFileData.id = fileId ?? DEFAULT_FOLDER_ID;
+  const mock: NewFileFunction = ({fileName, folderId}) => {
     newFileData.name = fileName;
     if (folderId) {
       newFileData.folderId = folderId;
@@ -20,8 +19,6 @@ const getNewFileMock = (parentId: FolderId): [ProjectFile, NewFileFunction] => {
 
   return [newFileData, mock];
 };
-
-const EXPECTED_NEXT_FILE_ID = '8';
 
 describe('openNewFilePrompt', function () {
   it('can successfully add a new file to root w/o validation file', async function () {
@@ -41,7 +38,6 @@ describe('openNewFilePrompt', function () {
       validationFile: undefined,
     });
 
-    expect(newFileData.id).toEqual(EXPECTED_NEXT_FILE_ID);
     expect(newFileData.name).toEqual(newFileName);
     expect(newFileData.folderId).toEqual(folderId);
 
@@ -64,7 +60,6 @@ describe('openNewFilePrompt', function () {
       validationFile: undefined,
     });
 
-    expect(newFileData.id).toEqual(EXPECTED_NEXT_FILE_ID);
     expect(newFileData.name).toEqual(newFileName);
     expect(newFileData.folderId).toEqual(folderId);
 
@@ -88,7 +83,6 @@ describe('openNewFilePrompt', function () {
       validationFile: undefined,
     });
 
-    expect(newFileData.id).toEqual(EXPECTED_NEXT_FILE_ID);
     expect(newFileData.name).toEqual(newFileName);
     expect(newFileData.folderId).toEqual(folderId);
 
@@ -112,7 +106,6 @@ describe('openNewFilePrompt', function () {
       validationFile,
     });
 
-    expect(newFileData.id).toEqual(String(Number(EXPECTED_NEXT_FILE_ID) + 1));
     expect(newFileData.name).toEqual(newFileName);
     expect(newFileData.folderId).toEqual(folderId);
