+++ apps/test/unit/codebridge/FileBrowser/prompts/openNewFolderPromptTest.ts
@@ -12,16 +12,13 @@ const getNewFolderMock = (
   parentId: FolderId
 ): [ProjectFolder, NewFolderFunction] => {
   const newFolderData = {} as ProjectFolder;
-  const mock: NewFolderFunction = ({parentId, folderName, folderId}) => {
+  const mock: NewFolderFunction = ({parentId, folderName}) => {
     newFolderData.parentId = parentId ?? DEFAULT_FOLDER_ID;
     newFolderData.name = folderName;
-    newFolderData.id = folderId;
   };
 
   return [newFolderData, mock];
 };
-
-const EXPECTED_NEXT_FOLDER_ID = '6';
 
 describe('openNewFolderPrompt', function () {
   it('can successfully add a new folder to root', async function () {
@@ -41,7 +38,6 @@ describe('openNewFolderPrompt', function () {
 
     expect(newFolderData.parentId).toEqual(parentId);
     expect(newFolderData.name).toEqual(newFolderName);
-    expect(newFolderData.id).toEqual(EXPECTED_NEXT_FOLDER_ID);
     expect(analyticsData.event).toEqual(EVENTS.CODEBRIDGE_NEW_FOLDER);
   });
 
@@ -61,7 +57,6 @@ describe('openNewFolderPrompt', function () {
 
     expect(newFolderData.parentId).toEqual(parentId);
     expect(newFolderData.name).toEqual(newFolderName);
-    expect(newFolderData.id).toEqual(EXPECTED_NEXT_FOLDER_ID);
     expect(analyticsData.event).toEqual(EVENTS.CODEBRIDGE_NEW_FOLDER);
   });
 
@@ -82,7 +77,6 @@ describe('openNewFolderPrompt', function () {
 
     expect(newFolderData.parentId).toEqual(parentId);
     expect(newFolderData.name).toEqual(newFolderName);
-    expect(newFolderData.id).toEqual(EXPECTED_NEXT_FOLDER_ID);
     expect(analyticsData.event).toEqual(EVENTS.CODEBRIDGE_NEW_SUBFOLDER);
   });
